requirements.txt
Django==2.1.1
upstox
python-dotenv
django-dotenv